refactor: move `feeds` package to `db.methods`
